Repositorio y Hreques impelemntado y testeado. Falta Vista de "mis publicaciones con estilo"

+++ src/main/java/ar/edu/unlam/tallerweb1/repositorios/RepositorioPropietario.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 
 public interface RepositorioPropietario {
-    List<Publicacion> obtenePublicacionesDelPropietario(Integer propietarioId);
+    List<Publicacion> obtenePublicacionesDelPropietario(Long propietarioId);
 }
